Move functions out of render fn. Create utils.
